+++ ui/app/src/store/modules/accounts.ts
@@ -1,4 +1,4 @@
-import { IAssetAmount, Network } from "@sifchain/sdk";
+import { IAsset, IAssetAmount, Network } from "@sifchain/sdk";
 import { useCore } from "../../hooks/useCore";
 import { Vuextra } from "../Vuextra";
 import { useChains } from "@/hooks/useChains";
@@ -109,7 +109,10 @@ export const accountStore = Vuextra.createStore({
       try {
         const address = await provider.connect(chain);
         self.setConnected({ network, connected: true });
-        self.setAddress({ network, address });
+
+        if (typeof address === "string") {
+          self.setAddress({ network, address });
+        }
       } catch (error) {
         console.error(network, "wallet connect error", error);
       } finally {
@@ -146,6 +149,28 @@ export const accountStore = Vuextra.createStore({
         self.state[chain.network].address,
       );
       self.setBalances({ network, balances });
+    },
+
+    async updateBalance(params: { network: Network; asset: IAsset }) {
+      const { network, asset } = params;
+      if (!self.state[network].connected) return;
+
+      const chain = useChains().get(network);
+      const provider = useCore().services.wallet.getPreferredProvider(chain);
+
+      const balance = await provider.fetchBalance(
+        chain,
+        self.state[chain.network].address,
+        asset.symbol,
+      );
+      const replacedBalances = self.state[network].balances.map((b) =>
+        b.symbol === balance.symbol ? balance : b,
+      );
+
+      self.setBalances({
+        network,
+        balances: replacedBalances.length ? replacedBalances : [balance],
+      });
     },
 
     async disconnect(network: Network) {